Week_07/G20200389010107/requirement.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
aniso8601==8.0.0
ansible==2.9.6
appnope==0.1.0
appscript==1.0.1
APScheduler==3.6.3
asn1crypto==1.0.1
astroid==2.3.1
astropy==3.2.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.2.0
Automat==0.8.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
boto3==1.11.0
botocore==1.14.0
Bottleneck==1.2.1
bz2file==0.98
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.8.3
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
distributed==2.5.2
docutils==0.15.2
dominate==2.5.1
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Moment==0.9.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
fsspec==0.5.2
future==0.17.1
gensim==3.8.0
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
graphviz==0.13.2
greenlet==0.4.15
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
incremental==17.5.0
inflect==4.1.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.42.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
keyring==18.0.0
kiwisolver==1.1.0
klayout==0.26.3
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
livereload==2.6.1
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nmap==0.0.1
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.17.2
numpydoc==0.9.1
objgraph==3.4.1
olefile==0.46
openpyxl==3.0.0
packaging==19.2
paddlepaddle-tiny==1.6.1
pandas==0.25.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
powerline-status==2.7
pretty-errors==1.2.16
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
PyDispatcher==2.0.5
pyecharts==1.7.1
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.4.2
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyPdf==1.13
pyrsistent==0.15.4
PySocks==1.7.1
pytesseract==0.3.3
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
pytest-runner==5.2
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
queuelib==1.5.0
requests==2.22.0
rope==0.14.0
ruamel-yaml==0.15.46
s3transfer==0.3.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Scrapy==1.6.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.85
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.9.0
snowballstemmer==2.0.0
snownlp==0.12.3
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
terminado==0.8.2
tesseract==0.1.3
testpath==0.4.2
toolz==0.10.0
torch==1.3.1
torchtext==0.5.0
torchvision==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
Twisted==19.10.0
tzlocal==2.0.0
unicodecsv==0.14.1
urllib3==1.24.2
visitor==0.1.3
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.11.2
WTForms==2.2.1
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
zope.interface==4.7.1
